Changed styling for login and register
