+++ frontend/src/App.css
@@ -17,7 +17,6 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background: rgba(0, 0, 0, 0.5);
   z-index: 1;
 }
