+++ frontend/src/App.js
@@ -1,15 +1,16 @@
 import React, { useState } from 'react';
 import './App.css';
 import Modal from './Modal';
-import LoginForm from './LoginForm';
-import SignUpForm from './SignUpForm';
+import LoginForm from './components/LoginForm';
+import SignUpForm from './components/SignUpForm';
 import backgroundVideo from './assets/animeclip1.mp4';
 import logo from './assets/FinalLogo.png';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Dashboard from './DashboardPage';
+import Dashboard from './components/DashboardPage';
+
 
 function App() {
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isModalOpen, setIsModalOpen] = useState(true);
   const [activeForm, setActiveForm] = useState('login');
 
   const toggleModal = () => setIsModalOpen(!isModalOpen);
@@ -28,21 +29,15 @@ function App() {
         </video>
         <div className="topbar">
           <img src={logo} alt="Logo" className="topbar-logo" />
-          <div className="buttons-container">
-            <button className="topbar-btn">Home</button>
-            <button className="topbar-btn">Services</button>
-            <button className="topbar-btn">Products</button>
-            <button className="get-started-btn" onClick={() => switchForm('login')}>GET STARTED</button>
-          </div>
         </div>
+        
         <Modal isOpen={isModalOpen} onClose={() => {
-          toggleModal();
           setActiveForm('login');
         }}>
           {activeForm === 'login' ? (
             <LoginForm 
               onClose={() => {
-                toggleModal();
+
                 setActiveForm('login');
               }} 
               onSwitchForm={() => switchForm('signup')} 
@@ -50,14 +45,12 @@ function App() {
           ) : (
             <SignUpForm 
               onClose={() => {
-                toggleModal();
-                setActiveForm('login');
+                setActiveForm('signup');
               }} 
               onSwitchBack={() => switchForm('login')}
             />
           )}
         </Modal>
-        {}
         <Routes>
           <Route path="/dashboard" element={<Dashboard />} />
           {}
